Tidied up previous commit
