+++ public/layouts/dashboard/dashboard-user.html
@@ -9,7 +9,7 @@
     </button>
 
     <div ng-if="showFeatureSwitch">
-        <wf-feature-switches></wf-feature-switches>
+        <wf-feature-switch></wf-feature-switch>
     </div>
 
     <ul class="dashboard-toolbar__dropdown dropdown-toolbar__list list-unstyled content-type-list--hidden" ng-class="{'content-type-list--hidden' : !showDropdown }">
